# apps/js/docs/swagger.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "BeliMudah API",
		"version": "1.0",
		"description": "E-commerce API for the BeliMudah storefront.",
		"contact": {
			"name": "Ryan Suhartanto",
			"url": "https://github.com/ryansuhartanto/koda-b8-backend",
			"email": "suhartanto@kekkon.nexus"
		},
		"license": {
			"name": "MIT"
		}
	},
	"servers": [
		{
			"url": "http://localhost:3002"
		}
	],
	"components": {
		"securitySchemes": {
			"BearerAuth": {
				"type": "http",
				"scheme": "bearer"
			}
		},
		"schemas": {
			"Problem": {
				"type": "object",
				"properties": {
					"title": {
						"type": "string"
					},
					"status": {
						"type": "integer"
					},
					"detail": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"status"
				]
			},
			"VariantOption": {
				"type": "object",
				"properties": {
					"option": {
						"type": "string"
					},
					"value": {
						"type": "string"
					}
				},
				"required": [
					"option",
					"value"
				]
			},
...
			"Category": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"icon": {
						"type": "string"
					},
					"img": {
						"type": "string"
					},
					"product_count": {
						"type": "integer"
					}
				},
				"required": [
					"id",
					"name",
					"product_count"
				]
			},
			"Brand": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"product_count": {
						"type": "integer"
					}
				},
				"required": [
					"id",
					"name",
					"product_count"
				]
			},
...
			"PaymentMethod": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"metadata": {
						"type": "object"
					}
				},
				"required": [
					"id",
					"metadata",
					"name"
				]
			},
			"CartItem": {
				"type": "object",
				"properties": {
					"id_variant": {
						"type": "string"
					},
					"id_product": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"variant_options": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/VariantOption"
						},
						"uniqueItems": false
					},
					"sku": {
						"type": "string"
					},
					"urls": {
						"type": "array",
						"items": {
							"type": "string"
						},
						"uniqueItems": false
					},
					"price_idr": {
						"type": "integer"
					},
					"original_price_idr": {
						"type": "integer"
					},
					"inventory": {
						"type": "integer"
					},
					"quantity": {
						"type": "integer"
					},
					"created_at": {
						"type": "string"
					}
				},
				"required": [
					"created_at",
					"id_product",
					"id_variant",
					"inventory",
					"name",
					"original_price_idr",
					"price_idr",
					"quantity"
				]
			},
			"Cart": {
				"type": "object",
				"properties": {
					"subtotal_idr": {
						"type": "integer"
					},
					"items": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/CartItem"
						},
						"uniqueItems": false
					}
				},
				"required": [
					"items",
					"subtotal_idr"
				]
			},
			"CartRequest": {
				"type": "object",
				"properties": {
					"id_variant": {
						"type": "string"
					},
					"quantity": {
						"type": "integer",
						"minimum": 1
					}
				},
				"required": [
					"id_variant",
					"quantity"
				]
			},
			"RegisterRequest": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string",
						"minLength": 8
					}
				},
				"required": [
					"email",
					"name",
					"password"
				]
			},
			"LoginRequest": {
				"type": "object",
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				]
			},
			"TokenResponse": {
				"type": "object",
				"properties": {
					"token": {
						"type": "string"
					}
				},
				"required": [
					"token"
				]
			},
			"Address": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"province": {
						"type": "string"
					},
					"postal_code": {
						"type": "string"
					},
					"is_default": {
						"type": "boolean"
					}
				},
				"required": [
					"address",
					"city",
					"id",
					"is_default",
					"label",
					"name",
					"phone",
					"postal_code",
					"province"
				]
			},
			"AddressRequest": {
				"type": "object",
				"properties": {
					"label": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"province": {
						"type": "string"
					},
					"postal_code": {
						"type": "string"
					},
					"is_default": {
						"type": "boolean"
					}
				},
				"required": [
					"address",
					"city",
					"label",
					"name",
					"phone",
					"postal_code",
					"province"
				]
			}
		}
	},
	"paths": {
...
		"/categories": {
			"get": {
				"summary": "List categories and how many products each holds",
				"tags": [
					"catalog"
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Category"
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					}
				}
			}
		},
		"/brands": {
			"get": {
				"summary": "List brands and how many products each holds",
				"tags": [
					"catalog"
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Brand"
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					}
				}
			}
		},
...
		"/payment-methods": {
			"get": {
				"summary": "List the payment methods an order can be placed with",
				"tags": [
					"catalog"
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/PaymentMethod"
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					}
				}
			}
		},
		"/me/cart": {
			"get": {
				"summary": "The caller's cart and its subtotal",
				"tags": [
					"cart"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Cart"
								}
							}
						}
					},
					"401": {
						"description": "Missing or invalid token",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					},
					"500": {
						"description": "Internal error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					}
				}
			},
			"post": {
				"summary": "Set the quantity of one cart line",
				"tags": [
					"cart"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"requestBody": {
					"description": "Line",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CartRequest",
								"summary": "body",
								"description": "Line"
							}
						}
					}
				},
				"responses": {
					"204": {
						"description": "No Content"
					},
					"400": {
						"description": "Invalid body",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					},
					"401": {
						"description": "Missing or invalid token",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					},
					"404": {
						"description": "No such variant",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					},
					"500": {
						"description": "Internal error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					}
				}
			}
		},
		"/me/cart/{id_variant}": {
			"delete": {
				"summary": "Remove one variant from the cart",
				"tags": [
					"cart"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id_variant",
						"required": true,
						"description": "Variant sqid",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"204": {
						"description": "No Content"
					},
					"401": {
						"description": "Missing or invalid token",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					},
					"404": {
						"description": "No such cart item",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					},
					"500": {
						"description": "Internal error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					}
				}
			}
		},
		"/auth/register": {
			"post": {
				"summary": "Register an account",
				"tags": [
					"auth"
				],
				"requestBody": {
					"description": "Credentials",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterRequest",
								"summary": "body",
								"description": "Credentials"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TokenResponse"
								}
							}
						}
					},
					"400": {
						"description": "Invalid body",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					},
					"409": {
						"description": "Email already registered",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					},
					"500": {
						"description": "Internal error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"summary": "Exchange credentials for a token",
				"tags": [
					"auth"
				],
				"requestBody": {
					"description": "Credentials",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequest",
								"summary": "body",
								"description": "Credentials"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TokenResponse"
								}
							}
						}
					},
					"400": {
						"description": "Invalid body",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					},
					"401": {
						"description": "Invalid credentials",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					},
					"500": {
						"description": "Internal error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					}
				}
			}
		},
		"/me/addresses": {
			"get": {
				"summary": "List the caller's addresses, default first",
				"tags": [
					"addresses"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Address"
									}
								}
							}
						}
					},
					"401": {
						"description": "Missing or invalid token",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					},
					"500": {
						"description": "Internal error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					}
				}
			},
			"post": {
				"summary": "Add an address to the caller's account",
				"tags": [
					"addresses"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"requestBody": {
					"description": "Address",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddressRequest",
								"summary": "body",
								"description": "Address"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Address"
								}
							}
						}
					},
					"400": {
						"description": "Invalid body",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					},
					"401": {
						"description": "Missing or invalid token",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					},
					"500": {
						"description": "Internal error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					}
				}
			}
		}
	}
}